+++ apps/web/src/apis/instance.ts
@@ -16,15 +16,16 @@ instance.interceptors.response.use(
     (res) => res,
     async (err) => {
         const originalRequest = err.config;
-
-        if (originalRequest._retry) {
+        if (!originalRequest || originalRequest._retry) {
             cookies.remove("accessToken");
             cookies.remove("refreshToken");
             window.location.replace("/login");
             return Promise.reject(err);
         }
 
-        if (err.response?.status !== 401) return Promise.reject(err);
+        if (err.response?.status !== 401) {
+            return Promise.reject(err);
+        }
 
         try {
             originalRequest._retry = true;
@@ -33,13 +34,12 @@ instance.interceptors.response.use(
                 throw new Error("리프레시 토큰이 없습니다.");
             }
 
-            const { data } = await axios.post("/auth/refresh", {
+            const { data } = await instance.post("/auth/refresh", {
                 token: refreshToken,
             });
 
             cookies.set("accessToken", data.accessToken);
-            instance.defaults.headers.common["Authorization"] =
-                `Bearer ${data.accessToken}`;
+            instance.defaults.headers.common.Authorization = `Bearer ${data.accessToken}`;
 
             return instance(originalRequest);
         } catch (error) {
